Abstract duplicated `sendTx` logic as `Signer` class
