+++ src/asset-burner.ts
@@ -1,21 +1,17 @@
-import { FungibleAssetValue, encodeCurrency, encodeSignedTx, signTx } from "@planetarium/tx";
-import { IHeadlessGraphQLClient } from "./interfaces/headless-graphql-client";
-import { Account } from "@planetarium/account";
-import { RecordView, encode } from "@planetarium/bencodex";
-import { additionalGasTxProperties } from "./tx";
+import { FungibleAssetValue, encodeCurrency } from "@planetarium/tx";
+import { RecordView } from "@planetarium/bencodex";
 import { IAssetBurner } from "./interfaces/asset-burner";
+import { Signer } from "./signer";
 
 export class AssetBurner implements IAssetBurner {
-    private readonly _client:  IHeadlessGraphQLClient;
-    private readonly _account: Account;
+    private readonly signer: Signer;
 
-    constructor(account: Account, client: IHeadlessGraphQLClient) {
-        this._account = account;
-        this._client = client;
+    constructor(signer: Signer) {
+        this.signer = signer;
     }
 
     async burn(amount: FungibleAssetValue, memo: string): Promise<string> {
-        const sender = (await this._account.getAddress()).toBytes();
+        const sender = (await this.signer.getAddress()).toBytes();
         const action = new RecordView(
             {
                 type_id: "burn_asset",
@@ -31,29 +27,6 @@ export class AssetBurner implements IAssetBurner {
             "text"
         );
 
-        return await this.sendTx(action);
-    }
-    
-    private async sendTx(action: RecordView): Promise<string> {
-        const address = await this._account.getAddress();
-        const nonce = BigInt(await this._client.getNextTxNonce(address.toHex()));
-        const genesisHash = Buffer.from(
-            await this._client.getGenesisHash(),
-            "hex"
-        );
-        
-        const unsignedTx = {
-            nonce,
-            genesisHash,
-            publicKey: (await this._account.getPublicKey()).toBytes("uncompressed"),
-            signer: address.toBytes(),
-            timestamp: new Date(),
-            updatedAddresses: new Set([]),
-            actions: [action],
-            ...additionalGasTxProperties,
-        };
-        
-        const tx = await signTx(unsignedTx, this._account);
-        return this._client.stageTransaction(Buffer.from(encode(encodeSignedTx(tx))).toString("hex"));
+        return await this.signer.sendTx(action);
     }
 }
