+++ src/asset-transfer.ts
@@ -1,21 +1,18 @@
 import { FungibleAssetValue, encodeCurrency, encodeSignedTx, signTx } from "@planetarium/tx";
 import { IAssetTransfer } from "./interfaces/asset-transfer";
-import { IHeadlessGraphQLClient } from "./interfaces/headless-graphql-client";
-import { Account, Address } from "@planetarium/account";
-import { RecordView, encode } from "@planetarium/bencodex";
-import { additionalGasTxProperties } from "./tx";
+import { Address } from "@planetarium/account";
+import { RecordView } from "@planetarium/bencodex";
+import { Signer } from "./signer";
 
 export class AssetTransfer implements IAssetTransfer {
-    private readonly _client:  IHeadlessGraphQLClient;
-    private readonly _account: Account;
+    private readonly signer: Signer;
 
-    constructor(account: Account, client: IHeadlessGraphQLClient) {
-        this._account = account;
-        this._client = client;
+    constructor(signer: Signer) {
+        this.signer = signer;
     }
 
     async transfer(recipient: Address, amount: FungibleAssetValue, memo: string): Promise<string> {
-        const sender = (await this._account.getAddress()).toBytes();
+        const sender = (await this.signer.getAddress()).toBytes();
         const action = new RecordView(
             {
                 type_id: "transfer_asset5",
@@ -30,29 +27,6 @@ export class AssetTransfer implements IAssetTransfer {
             "text"
         );
 
-        return await this.sendTx(action);
-    }
-    
-    private async sendTx(action: RecordView): Promise<string> {
-        const address = await this._account.getAddress();
-        const nonce = BigInt(await this._client.getNextTxNonce(address.toHex()));
-        const genesisHash = Buffer.from(
-            await this._client.getGenesisHash(),
-            "hex"
-        );
-        
-        const unsignedTx = {
-            nonce,
-            genesisHash,
-            publicKey: (await this._account.getPublicKey()).toBytes("uncompressed"),
-            signer: address.toBytes(),
-            timestamp: new Date(),
-            updatedAddresses: new Set([]),
-            actions: [action],
-            ...additionalGasTxProperties,
-        };
-        
-        const tx = await signTx(unsignedTx, this._account);
-        return this._client.stageTransaction(Buffer.from(encode(encodeSignedTx(tx))).toString("hex"));
+        return await this.signer.sendTx(action);
     }
 }
