+++ src/minter.ts
@@ -1,18 +1,15 @@
-import { Account, Address } from "@planetarium/account";
-import { encodeCurrency, encodeSignedTx, signTx } from "@planetarium/tx";
+import { Address } from "@planetarium/account";
+import { encodeCurrency } from "@planetarium/tx";
 import { IFungibleAssetValues, IFungibleItems, IMinter } from "./interfaces/minter";
-import { HeadlessGraphQLClient } from "./headless-graphql-client";
-import { RecordView, Value, encode } from "@planetarium/bencodex";
-import { additionalGasTxProperties } from "./tx";
+import { RecordView, Value } from "@planetarium/bencodex";
+import { Signer } from "./signer";
 
 
 export class Minter implements IMinter {
-    private readonly _account: Account;
-    private readonly _client: HeadlessGraphQLClient;
+    private readonly signer: Signer;
     
-    constructor(account: Account, client: HeadlessGraphQLClient) {
-        this._account = account;
-        this._client = client;
+    constructor(signer: Signer) {
+        this.signer = signer;
     }
     
     async mintAssets(assets: [IFungibleAssetValues | IFungibleItems]): Promise<string> {
@@ -24,34 +21,11 @@ export class Minter implements IMinter {
             "text"
         );
 
-        return await this.sendTx(action);
+        return await this.signer.sendTx(action);
     }
     
     getMinterAddress(): Promise<Address> {
-        return this._account.getAddress();
-    }
-
-    private async sendTx(action: RecordView): Promise<string> {
-        const minter = await this.getMinterAddress();
-        const nonce = BigInt(await this._client.getNextTxNonce(minter.toHex()));
-        const genesisHash = Buffer.from(
-            await this._client.getGenesisHash(),
-            "hex"
-        );
-        
-        const unsignedTx = {
-            nonce,
-            genesisHash,
-            publicKey: (await this._account.getPublicKey()).toBytes("uncompressed"),
-            signer: minter.toBytes(),
-            timestamp: new Date(),
-            updatedAddresses: new Set([]),
-            actions: [action],
-            ...additionalGasTxProperties,
-        };
-        
-        const tx = await signTx(unsignedTx, this._account);
-        return this._client.stageTransaction(Buffer.from(encode(encodeSignedTx(tx))).toString("hex"));
+        return this.signer.getAddress();
     }
 }
